- Updated for Dofus 2.4.2.50422.1 - GameServer now handles ChangeMapMessage, GameMapMovementConfirmMessage, GameMapMovementRequestMessage, and basic AdminCommandMessage/AdminQuietCommandMessage (actually just quiet teleportation), but currently there isn't pathfinding system - GameServer now sends the character stats list to the client - Added basic map handling system - GameServer architecture entierely finished - Some refactoring and minor corrections - And I may have forgotten many other things!

+++ Shared/DofusModel.h
@@ -24,7 +24,7 @@ class DofusModel
 protected:
 	ByteBuffer m_buffer;
 public:
-	virtual uint32 GetProtocol() const = 0;
+	virtual uint16 GetProtocol() const = 0;
 
 	operator ByteBuffer() const
 	{ return m_buffer; }
